requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.3

# Configuration Management
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0
